Sample points from geographies and write coordinates to xml

+++ pam/write.py
@@ -119,8 +119,8 @@ def write_matsim_plans(population, location, comment=None):
                 if isinstance(component, Activity):
                     et.SubElement(plan_xml, 'act', {
                         'type': component.act,
-                        'x': str(int(component.location.loc.x)),
-                        'y': str(int(component.location.loc.y)),
+                        'x': str(component.location.loc.x),
+                        'y': str(component.location.loc.y),
                         'end_time': dttm(component.end_time)
                     }
                                   )
@@ -132,8 +132,8 @@ def write_matsim_plans(population, location, comment=None):
             component = person[-1]  # write the last activity without an end time
             et.SubElement(plan_xml, 'act', {
                 'type': component.act,
-                'x': str(int(component.location.loc.x)),
-                'y': str(int(component.location.loc.y)),
+                'x': str(float(component.location.loc.x)),
+                'y': str(float(component.location.loc.y)),
             }
                           )
